feat: video, debug, list, disable tools
