+++ examples/savetext.md
@@ -1,0 +1,16 @@
+Prompt: save a short poem about postman collections.
+
+Thought: Let me create the poem first.
+Final Answer:
+Postman collections, oh so grand,
+Organizing requests, just as planned.
+APIs tested, endpoints explored,
+With every response, our knowledge soared.
+
+In folders nested, we keep them neat,
+Each environment variable, a special treat.
+From local to cloud, we share with ease,
+Collaboration and teamwork, sure to please.
+
+Postman collections, our trusty guide,
+Through the world of APIs, we stride.
