+++ package.json
@@ -23,7 +23,7 @@
     "meta-fetcher": "^3.1.1",
     "open": "^9.1.0",
     "svg2png-wasm": "^1.3.4",
-    "terminal-image": "^2.0.0",
+    "termview": "^1.0.0",
     "turndown": "github:mixmark-io/turndown",
     "turndown-plugin-gfm": "^1.0.2",
     "uuid": "^9.0.0",
